cleanup in benefit schema and moved filtering of benefits to compensatins
